style(tests): move the new reporter tests to org.testng and settle the vocabulary
The tests this branch adds sat under test.*, where the tree has been moving to
org.testng.<area> for a while: org.testng.reporters.issue2725 and
org.testng.groups.issue2232 are the shape. JqReportMemoryTest, its launcher
and its sample move to org.testng.reporters.issue1259, and FileStringBufferTest
joins XmlCharFilteringWriterTest in org.testng.reporters, beside the class it
covers. That package is @NullMarked, so the test that appends null -- which is
the whole point of it -- carries @SuppressWarnings("NullAway"), the spelling
XmlTestTest already uses for a deliberate null.

Six identifiers reached prose that Vale reads: charset, dereferences,
getContent, maxTime, toWriter and addBuffer. accept.txt already carries several
hundred of those, dereference and dereferencing among them, so the inflection
was the thing missing rather than the practice. One filler word goes with them.

writingStyleCheckChanges is clean on every line this branch touched. Three
findings remain on lines it did not.
