+++ .vale/styles/config/vocabularies/TestNG/accept.txt
@@ -87,10 +87,10 @@ XPaths
 Xerces
 Xml
 Yaml
-aString
 accessor
 accessors
 actualArguments
+addBuffer
 addEdge
 addListener
 afterClass
@@ -108,6 +108,7 @@ argv
 arity
 asIterable
 assertEquals
+aString
 async
 autostyle
 backgrounded
@@ -122,11 +123,12 @@ boolean
 bugtracker
 cbeust
 cedric
+charset
 cla
 classListeners
-className
 classloader
 classname
+className
 classpath
 clazz
 cli
@@ -136,19 +138,19 @@ colspan
 combinatorics
 commandline
 confFail
-confSkip
 config
-configMethod
 configfailurepolicy
+configMethod
 configurers
+confSkip
 containsExactly
 containsOnly
 cquezel
 cssClasses
 ctor
 ctx
-dataProvider
 dataprovider
+dataProvider
 deduplicated
 deduplicates
 defaultComment
@@ -156,11 +158,12 @@ defaultListeners
 dependsOn
 dequeues
 dereference
+dereferences
 dereferencing
 deserialize
 desynchronize
-docType
 doctype
+docType
 doctypes
 dotDefinition
 dtd
@@ -178,12 +181,12 @@ existingMethod
 expectedExceptions
 expirationDate
 factoryClass
-failMessage
 failedTests
+failMessage
 failover
 failsafe
-failurePolicy
 failure_but_within_successPercentage
+failurePolicy
 fedInstance
 filePath
 finalizer
@@ -198,14 +201,15 @@ freeNodes
 freynaud
 gc
 generateReport
+getContent
 getObjects
 getters
 grep
 groupRegexp
-hasNext
 hashCode
 hashcode
 hashmaps
+hasNext
 hhmm
 href
 html
@@ -213,11 +217,11 @@ https
 icb
 icl
 ie
-inFormat
 include
 includedGroups
 includedMethods
 independentNode
+inFormat
 initializer
 injectables
 injectionMethod
@@ -250,8 +254,8 @@ krmahadevan
 listenerClass
 listenerInstance
 listenerMethod
-loadClasses
 loadability
+loadClasses
 localName
 lookups
 lowlink
@@ -268,6 +272,7 @@ m_testListeners
 m_verbose
 maintainer
 maxDepth
+maxTime
 memoize
 memoized
 memoizing
@@ -308,27 +313,27 @@ osgi
 outDir
 outFormat
 outMethod
+outputdir
 outputDir
 outputDirectory
 outputFolder
-outputdir
-pClass
 packageName
-parallelMode
 parallelization
+parallelMode
 param
-paramIndex
-paramValues
+parameterless
 parameterName
 parameterNames
 parameterType
 parameterTypes
 parameterValues
-parameterless
+paramIndex
 params
+paramValues
 parser
 passedResult
 passedTests
+pClass
 pluggable
 precompiled
 prepended
@@ -352,7 +357,6 @@ rewraps
 rootNode
 runInfo
 runningMethods
-sName
 scm
 secondParameter
 serializable
@@ -362,6 +366,7 @@ sha
 shas
 skippedTests
 snakeyaml
+sName
 snapshotted
 snapshotting
 stacktrace
@@ -383,22 +388,22 @@ superset
 tagName
 tearDown
 teardown
+testcase
 testClass
 testClasses
 testConstructor
+testcontext
 testContext
 testInstances
 testMethod
 testMethods
-testNGMethod
 testName
 testNames
-testResult
-testcase
-testcontext
 testnames
 testng
+testNGMethod
 testngMethod
+testResult
 testsuite
 testsuites
 thirdParameter
@@ -416,12 +421,13 @@ tm
 tng
 toDot
 toHtml
-toString
-toYaml
 toolchain
 toolchains
 topologicalSort
+toString
 totalRuns
+toWriter
+toYaml
 triaged
 triaging
 uncompilable
@@ -446,12 +452,12 @@ xmlBuffer
 xmlClass
 xmlInclude
 xmlName
+xmlns
 xmlPath
+xmls
 xmlSuite
 xmlSuites
 xmlTest
-xmlns
-xmls
 xsb
 xy
 yaml
