+++ CHANGES.txt
@@ -1,5 +1,5 @@
 Current (7.13.0)
-Fixed: org.testng.reporters.FileStringBuffer no longer throws a NullPointerException when the first thing appended to it is a string of 100 000 characters or more. A string that size bypasses the in-memory builder and is written to the temporary file directly, and the file is created as the builder is flushed -- but flushToFile answers early when the builder is empty, which it is when nothing has been appended yet, so that path went straight on to write to a file it had just decided not to create. Appending anything at all first, however short, was enough to hide it. The file is now created where it is needed rather than as a side effect of a flush, and a buffer whose file cannot be created says so with an IllegalStateException naming it instead of failing as a null one frame later, which is how FileStringBuffer.toString already reports the same impossibility. test.FileStringBufferTest, which had asserted nothing since it was written in 2013 and was listed in no suite, now covers both readers -- toString and toWriter -- below the flush threshold, across it, and on either side of that direct-to-file size. Two silent failures in the same class go with it: a flush that could not write used to replace the in-memory builder anyway, so the characters it still held vanished from the middle of the document with nothing raised, and toWriter logged what toString has always thrown for, so a spilled buffer whose temporary file had gone answered an empty document instead. Both now raise IllegalStateException, which is what temporaryFile already did for the same class of fault. And the temporary file is written and read as UTF-8 rather than through the platform default charset: a character that charset could not encode was lost at spill time, before the explicitly UTF-8 page writer ever saw it, so a supplementary character in a report reached index.html as a question mark on a JDK whose default is not UTF-8 (Julien Herr)
+Fixed: org.testng.reporters.FileStringBuffer no longer throws a NullPointerException when the first thing appended to it is a string of 100 000 characters or more. A string that size bypasses the in-memory builder and is written to the temporary file directly, and the file is created as the builder is flushed -- but flushToFile answers early when the builder is empty, which it is when nothing has been appended yet, so that path went straight on to write to a file it had just decided not to create. Appending anything at all first, however short, was enough to hide it. The file is now created where it is needed rather than as a side effect of a flush, and a buffer whose file cannot be created says so with an IllegalStateException naming it instead of failing as a null one frame later, which is how FileStringBuffer.toString already reports the same impossibility. org.testng.reporters.FileStringBufferTest, which had asserted nothing since it was written in 2013 and was listed in no suite, now covers both readers -- toString and toWriter -- below the flush threshold, across it, and on either side of that direct-to-file size. Two silent failures in the same class go with it: a flush that could not write used to replace the in-memory builder anyway, so the characters it still held vanished from the middle of the document with nothing raised, and toWriter logged what toString has always thrown for, so a spilled buffer whose temporary file had gone answered an empty document instead. Both now raise IllegalStateException, which is what temporaryFile already did for the same class of fault. And the temporary file is written and read as UTF-8 rather than through the platform default charset: a character that charset could not encode was lost at spill time, before the explicitly UTF-8 page writer ever saw it, so a supplementary character in a report reached index.html as a question mark on a JDK whose default is not UTF-8 (Julien Herr)
 Fixed: GITHUB-1259, GITHUB-2334: index.html is no longer built by holding one panel of it in memory at a time. Every panel of the default HTML report writes into an XMLStringBuffer of its own -- backed, like the page it feeds, by a buffer that spills to a temporary file past 100 000 characters -- and used to hand that buffer to the page as a String, which read the whole file back and copied it once more for the regular expression that drops the characters XML 1.0 does not allow. So a buffer whose reason to exist is not being held in memory was held in memory twice over, once per panel, and the largest of them grows with the result count: GITHUB-2334 ran out of heap in NavigatorPanel.generateMethodList listing the passed methods of a run of 500 000, GITHUB-1259 in BaseMultiSuitePanel, both of them under a summary saying every test of the run had passed. The panels now stream their buffer into the page through the new XMLStringBuffer.addBuffer, which drops the same characters a slice at a time, and TimesPanel writes its table a row at a time rather than building the whole of it as one String. On a run of 20 000 results whose report is 39 MB, generating it needed a 64 MB heap and now needs 28 MB; below that the limit is what the run itself retains, which is GITHUB-1979 and unchanged here. The page is identical, the order of what it lists included. org.testng.reporters.XMLStringBuffer.addBuffer(XMLStringBuffer) is new, and the panels of org.testng.reporters.jq write into a buffer BaseMultiSuitePanel owns, through a new package-private writeContent, rather than each building one and answering its String. BaseMultiSuitePanel.getContent(ISuite, XMLStringBuffer) stays where it is -- concrete now rather than abstract -- for callers outside the report, and is the one place left that materialises a panel. BaseMultiSuitePanel.getContent(ISuite, XMLStringBuffer) is final: generate no longer calls it, so an override outside the package decided nothing and would have been skipped without a word. TimesPanel no longer adds each result's duration into its running total on every call, which made a second call to getContent report the suite as having run twice as long (Julien Herr)
 Fixed: GITHUB-1031: A data provider that supplies the wrong number or type of arguments now reports that mismatch. Count errors state the expected and actual counts. Type errors state that the types do not match. The matcher no longer claims the test method has no parameters (Burak Kalaycı)
 Fixed: GITHUB-3121: EmailableReporter2 now appends each test invocation's non-blank ITest custom name to the Java method name in the summary table and scenario headings. Configuration methods keep the Java method name (Burak Kalaycı)
