+++ testng-core/src/test/java/org/testng/reporters/issue1259/LargeReportSample.java
@@ -1,4 +1,4 @@
-package test.reports.issue1259;
+package org.testng.reporters.issue1259;
 
 import java.util.Iterator;
 import java.util.stream.IntStream;
